+++ src/WorldCupBets.Application/Features/Bets/PlaceChampionBetResultDto.cs
@@ -2,6 +2,6 @@ namespace WorldCupBets.Application.Features.Bets;
 
 public sealed record PlaceChampionBetResultDto(
     string TeamName,
-    int StakeAmountCc,
-    int RemainingBalanceCc,
+    decimal StakeAmountCc,
+    decimal RemainingBalanceCc,
     DateTime PlacedAtUtc);
