+++ src/WorldCupBets.Application/Features/Bets/PlaceMatchBetResultDto.cs
@@ -3,6 +3,6 @@ namespace WorldCupBets.Application.Features.Bets;
 public sealed record PlaceMatchBetResultDto(
     int MatchId,
     string Selection,
-    int StakeAmountCc,
-    int RemainingBalanceCc,
+    decimal StakeAmountCc,
+    decimal RemainingBalanceCc,
     DateTime PlacedAtUtc);
