+++ tests/WorldCupBets.Application.Tests/MatchRulesTests.cs
@@ -159,7 +159,7 @@ public sealed class MatchRulesTests
             return Task.FromResult<IReadOnlyList<MatchBet>>(matchBets.Where(matchBet => matchBet.MatchId == matchId).ToArray());
         }
 
-        public Task<IReadOnlyDictionary<int, int>> ListPendingStakeAmountsByUserAsync(CancellationToken cancellationToken = default)
+        public Task<IReadOnlyDictionary<int, decimal>> ListPendingStakeAmountsByUserAsync(CancellationToken cancellationToken = default)
         {
             throw new NotSupportedException();
         }
